Check whether synsights-metrics is enabled in `inv.applications`
With https://github.com/projectsyn/commodore/issues/223 we're moving to
using the reclass `applications` array to define which components are to
be included, instead of directly including component classes in the
hierarchy.

Therefore the component needs to check in `inv.applications` instead of
`inv.classes` whether synsights-metrics is enabled to decide whether to
deploy the ArgoCD monitoring.