Fix label selectors for ArgoCD servicemonitors
We renamed the ServiceMonitors in #129 without verifying that the
renamed service monitors still discover any endpoints.

We need to split the ServiceMonitor name from the name used for the
label selector to ensure that we can use an object name which doesn't
collide with the operator-managed name, while still picking up the
endpoints.