feat: Allow creation of ArgoCDs and AppProjects
While this component already managed the ArgoCD operator and the Syn
ArgoCD instance, additional ArgoCD instances could only be deployed as
adhoc-manifests. While this works, it is not possible to leverage Syn
features such as hierarchical configuration, resulting in quite some
duplication.

This commit introduces the new parameters `argocds` and `projects`,
which allow the creation of `ArgoCD` and `AppProject` instances
respectively.

In general, the component tries to be as un-intrusive as possible, so we
only set minimal default values:

For `ArgoCD`: spec.applicationInstanceLabelKey` - if this value collides
for two ArgoCD instances on the same cluster, ArgoCD cannot properly
determine which resources belong to "itself", and might try to prune
resources managed by other ArgoCD instances.

Signed-off-by: Manuel Hutter <manuel@hutter.io>