Nest AppProjects under ArgoCD
As mentioned in the code review, AppProject resources created usig this
component usually only exist within the context of an ArgoCD instance
ALSO deployed using this component, so it makes sense to nest these
resources.

As a side-effect, this allows me to use `instances` as the parameter
key, which I prefer :-)

Additionally we now validate that no two instances are deployed to the
same namespace.

Signed-off-by: Manuel Hutter <manuel@hutter.io>