Deploy cronjob which periodically refreshes the `syn-argocd-tls` secret
Unfortunately, the argocd-operator currently doesn't refresh the
certificate stored in secret `syn-argocd-tls` even when the certificate
is expired or expires soon.

To circumvent the certificate expiring (the lifetime is hardcoded to 1
year), we deploy a CronJob which deletes the `syn-argocd-tls` secret
every 4 months to force the operator to recreate it with a new
certificate.