Add support for https catalog repo
Because component-argocd can't use Vault references directly (due to
needing to be able to be synced to setup the Vault integration during
cluster bootstrap), we simply provide a parameter which contains the
name of an externally managed secret in the `syn` namespace.

This secret must have keys `username` and `password` which will be
configured as the credentials for the cluster catalog if the catalog URL
starts with `https://`.