Disable pruning for automated sync of root app
This ensures that we'll be able to roll out support for multiple root
apps without having to disable auto-sync for the default root app before
the roll out, since we already have some configuration that will
reassign existing component instances to another root app.

Later, we'll want to keep this setting, so we can move component
instances between different root apps without having to manually disable
auto-sync for the root apps on all clusters that use the component
that's being moved.