Add dynamic `syn_team` label to alerts
We use a small go template to extract the value for `syn_team` from the
`project` label of the metric so we can correctly label each individual
alert instance. This should allow us to group ArgoCD alerts by
`syn_team` (which is already well-defined for alert routing) to ensure
that alerts for each AppProject are routed to the responsible team.