Implement custom health check for nmstate.io `NodeNetworkConfigurationPolicy`
By implementing a custom health check for NNCPs, we can use ArgoCD sync
waves to only delay creation of resources which depend on the
configuration that's deployed on one or more nodes via NNCPs until the
NNCP contents are actually live on all targeted nodes.

For example, when using NNCPs to dynamically create dummy interfaces
with egress IPs for Cilium egress gateway policies, we need to delay the
creation of the egress policy until the dummy interfaces are actually
created.